// frontend/src/lib/openapi.d.ts
export interface components {
    schemas: {
// ...
        CategoryTreeOut: {
            /** Id */
            id: number;
            /** Parent Id */
            parent_id?: number | null;
            /** Name */
            name: string;
            /** Slug */
            slug: string;
            /** Description */
            description?: string | null;
            /** Position */
            position: number;
            /**
             * Images
             * @default []
             */
            images: components["schemas"]["ImageOut"][];
            /**
             * Children
             * @default []
             */
            children: components["schemas"]["CategoryTreeOut"][];
        };
// ...
        ProductOut: {
            /** Id */
            id: number;
            /** Category Id */
            category_id: number;
            /** Name */
            name: string;
            /** Slug */
            slug: string;
            /** Description */
            description?: string | null;
            /** Price */
            price: string;
            /** Sku */
            sku?: string | null;
            /** In Stock */
            in_stock: boolean;
            /**
             * Images
             * @default []
             */
            images: components["schemas"]["ImageOut"][];
        };
// ...
        UserOut: {
            /** Id */
            id: number;
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Is Active */
            is_active: boolean;
            /** Is Superuser */
            is_superuser: boolean;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
            /** Input */
            input?: unknown;
            /** Context */
            ctx?: Record<string, never>;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}